+++ go/test/endtoend/onlineddl/ghost/onlineddl_ghost_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package onlineddl
+package ghost
 
 import (
 	"context"
