+++ go/test/endtoend/onlineddl/revert/onlineddl_revert_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package onlineddl
+package revert
 
 import (
 	"context"
