+++ go/test/endtoend/onlineddl/vrepl/onlineddl_vrepl_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package onlineddl
+package vrepl
 
 import (
 	"context"
